feat: refactored functions into utilities
